python/requirements.txt
cffi==1.15.0
click==8.1.3
cryptography==36.0.1
ecdsa==0.17.0
Flask==2.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
pycryptodome==3.15.0
six==1.15.0
Werkzeug==2.1.2
zipp==3.8.0
